+++ test/cases/phase2/let-cast/explicit-as/expected.c
@@ -4,7 +4,7 @@ int main(void) {
     TSC_INIT();
     int64_t a = 1LL;
     uint32_t b = 2U;
-    const double c = (double)(a + (int64_t)b);
+    const double c = (double)((int64_t)((uint64_t)a + (uint64_t)(int64_t)b));
     printf("%g\n", (double)(c));
     return 0;
 }
